+++ src/ratroyale/backend/game_manager.py
@@ -1,9 +1,7 @@
 import math
 from random import shuffle
 
-from ratroyale.entities.rodent import Rodent
-from ratroyale.entity import Entity, SkillResult
-
+from .entity import Entity, SkillResult
 from .hexagon import OddRCoord
 from .player_info.player_info import PlayerInfo
 from .player_info.squeak import Squeak
